refactor(model): remove stale packing module
