+++ molt/trainer/workers/critic_actor.py
@@ -42,9 +42,9 @@ from torch.utils.data import DataLoader
 from tqdm import tqdm
 
 from molt.models import Critic, ValueLoss
+from molt.models.utils import unshard_dtensor
 from molt.trainer.algorithm.experience import Experience, get_model_parallel_size
 from molt.trainer.fsdp import FsdpStrategy
-from molt.trainer.fsdp.packing import unshard_dtensor
 from molt.utils import get_tokenizer
 from molt.utils.distributed_util import torch_dist_barrier_and_cuda_sync
 from molt.utils.logging_utils import init_logger
